# docs/openapi.yaml
...
openapi: "3.0.1"
info:
  title: "API Системы Управления Банковскими Картами"
  description: "REST API для управления банковскими картами с функциями:\n\n- Ауте\
    нтификация и авторизация пользователей\n- Управление картами (создание, просмо\
    тр, блокировка)\n- Переводы между картами\n- Администрирование системы\n\n**Ро\
    ли пользователей:**\n- USER: базовые операции с собственными картами\n- ADMIN:\
    \ полный доступ к системе\n\n**Безопасность:**\n- JWT токены для аутентификаци\
    и\n- Шифрование номеров карт\n- Маскирование чувствительных данных\n"
  contact:
    name: "Abu"
    url: "https://github.com/DesBasito"
    email: "out1of1mind1exception@gmail.com"
  version: "1.0.0"
servers:
- url: "http://localhost:8080/api/v1"
  description: "Локальный сервер разработки"
security:
- Bearer Authentication: []
...
paths:
  /users:
    get:
      tags:
      - "Управление пользователями"
      summary: "Получить всех пользователей"
      description: "Получение списка всех пользователей с пагинацией"
      operationId: "getAllUsers"
      parameters:
      - name: "pageable"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageUserDto"
      security:
      - Bearer Authentication: []
    post:
      tags:
      - "Управление пользователями"
      summary: "Создать пользователя"
      description: "Создание нового пользователя администратором"
      operationId: "createUser"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
      security:
      - Bearer Authentication: []
  /users/{id}/toggle-status:
    post:
      tags:
      - "Управление пользователями"
      summary: "Заблокировать/разблокировать пользователя"
      description: "Изменение статуса активности пользователя"
      operationId: "toggleUserStatus"
      parameters:
      - name: "id"
        in: "path"
        description: "ID пользователя"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
      security:
      - Bearer Authentication: []
...
  /auth/sign-up:
    post:
      tags:
      - "Аутентификация"
      summary: "Регистрация пользователя"
      operationId: "signUp"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
        required: true
      responses:
        default:
          description: "Возвращает токен после регистрации"
          content:
            '*/*':
              schema:
                type: "string"
                enum:
                - "100 CONTINUE"
                - "101 SWITCHING_PROTOCOLS"
                - "102 PROCESSING"
                - "103 EARLY_HINTS"
                - "103 CHECKPOINT"
                - "200 OK"
                - "201 CREATED"
                - "202 ACCEPTED"
                - "203 NON_AUTHORITATIVE_INFORMATION"
                - "204 NO_CONTENT"
                - "205 RESET_CONTENT"
                - "206 PARTIAL_CONTENT"
                - "207 MULTI_STATUS"
                - "208 ALREADY_REPORTED"
                - "226 IM_USED"
                - "300 MULTIPLE_CHOICES"
                - "301 MOVED_PERMANENTLY"
                - "302 FOUND"
                - "302 MOVED_TEMPORARILY"
                - "303 SEE_OTHER"
                - "304 NOT_MODIFIED"
                - "305 USE_PROXY"
                - "307 TEMPORARY_REDIRECT"
                - "308 PERMANENT_REDIRECT"
                - "400 BAD_REQUEST"
                - "401 UNAUTHORIZED"
                - "402 PAYMENT_REQUIRED"
                - "403 FORBIDDEN"
                - "404 NOT_FOUND"
                - "405 METHOD_NOT_ALLOWED"
                - "406 NOT_ACCEPTABLE"
                - "407 PROXY_AUTHENTICATION_REQUIRED"
                - "408 REQUEST_TIMEOUT"
                - "409 CONFLICT"
                - "410 GONE"
                - "411 LENGTH_REQUIRED"
                - "412 PRECONDITION_FAILED"
                - "413 PAYLOAD_TOO_LARGE"
                - "413 REQUEST_ENTITY_TOO_LARGE"
                - "414 URI_TOO_LONG"
                - "414 REQUEST_URI_TOO_LONG"
                - "415 UNSUPPORTED_MEDIA_TYPE"
                - "416 REQUESTED_RANGE_NOT_SATISFIABLE"
                - "417 EXPECTATION_FAILED"
                - "418 I_AM_A_TEAPOT"
                - "419 INSUFFICIENT_SPACE_ON_RESOURCE"
                - "420 METHOD_FAILURE"
                - "421 DESTINATION_LOCKED"
                - "422 UNPROCESSABLE_ENTITY"
                - "423 LOCKED"
                - "424 FAILED_DEPENDENCY"
                - "425 TOO_EARLY"
                - "426 UPGRADE_REQUIRED"
                - "428 PRECONDITION_REQUIRED"
                - "429 TOO_MANY_REQUESTS"
                - "431 REQUEST_HEADER_FIELDS_TOO_LARGE"
                - "451 UNAVAILABLE_FOR_LEGAL_REASONS"
                - "500 INTERNAL_SERVER_ERROR"
                - "501 NOT_IMPLEMENTED"
                - "502 BAD_GATEWAY"
                - "503 SERVICE_UNAVAILABLE"
                - "504 GATEWAY_TIMEOUT"
                - "505 HTTP_VERSION_NOT_SUPPORTED"
                - "506 VARIANT_ALSO_NEGOTIATES"
                - "507 INSUFFICIENT_STORAGE"
                - "508 LOOP_DETECTED"
                - "509 BANDWIDTH_LIMIT_EXCEEDED"
                - "510 NOT_EXTENDED"
                - "511 NETWORK_AUTHENTICATION_REQUIRED"
  /auth/sign-in:
    post:
      tags:
      - "Аутентификация"
      summary: "Авторизация пользователя"
      operationId: "signIn"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInRequest"
        required: true
      responses:
        default:
          description: "Возвращает токен после авторизации, рефреш устанавливает в\
            \ куку"
          content:
            '*/*':
              schema:
                type: "string"
  /auth/refresh:
    post:
      tags:
      - "Аутентификация"
      summary: "Обновления access токена для JWT"
      operationId: "refreshToken"
      parameters:
      - name: "refreshToken"
        in: "cookie"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        default:
          description: "Возвращает новый обновленный токен, рефреш устанавливает в\
            \ куку"
          content:
            '*/*':
              schema:
                type: "string"
  /users/{id}:
    get:
      tags:
      - "Управление пользователями"
      summary: "Получить пользователя по ID"
      description: "Получение подробной информации о пользователе"
      operationId: "getUser"
      parameters:
      - name: "id"
        in: "path"
        description: "ID пользователя"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "404":
          description: "Пользователь не найден"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
        "200":
          description: "Информация о пользователе получена"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDto"
      security:
      - Bearer Authentication: []
    delete:
      tags:
      - "Управление пользователями"
      summary: "Удалить пользователя"
      description: "Удаление пользователя из системы"
      operationId: "deleteUser"
      parameters:
      - name: "id"
        in: "path"
        description: "ID пользователя"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "Пользователь удален"
          content:
            '*/*':
              schema:
                type: "string"
        "400":
          description: "Нельзя удалить пользователя с активными картами"
          content:
            '*/*':
              schema:
                type: "string"
      security:
      - Bearer Authentication: []
...
components:
  schemas:
    SignUpRequest:
      required:
      - "name"
      - "password"
      - "phoneNumber"
      - "surname"
      type: "object"
      properties:
        name:
          type: "string"
          description: "Имя пользователя"
          example: "Сергей"
        surname:
          type: "string"
          description: "Фамилия пользователя"
          example: "Сергеев"
        middleName:
          type: "string"
          description: "Отчество пользователя"
          example: "Сергеевич"
        phoneNumber:
          pattern: "^\\+7\\([0-9]{3}\\)[0-9]{3}[0-9]{4}$"
          type: "string"
          description: "Мобильный телефон пользователя"
          example: "+7(900)1234567"
        password:
          maxLength: 11
          minLength: 5
          type: "string"
          description: "Пароль для пользователя"
          example: "password"
        fullName:
          type: "string"
      description: "Запрос на регистрацию"
...
    CardDto:
      required:
      - "balance"
      - "cardNumber"
      - "createdAt"
      - "expiryDate"
      - "ownerName"
      - "status"
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        cardNumber:
          type: "string"
        ownerName:
          type: "string"
        ownerId:
          type: "integer"
          format: "int64"
        expiryDate:
          type: "string"
          format: "date"
        status:
          type: "string"
        type:
          type: "string"
        balance:
          type: "number"
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
...
    SignInRequest:
      required:
      - "password"
      - "phoneNumber"
      type: "object"
      properties:
        phoneNumber:
          pattern: "^\\+7\\([0-9]{3}\\)[0-9]{3}[0-9]{4}$"
          type: "string"
          description: "Номер телефона пользователя"
          example: "777010203"
        password:
          type: "string"
          description: "Пароль"
          example: "my_1secret1_password"
      description: "Запрос на аутентификацию"
    Pageable:
      type: "object"
      properties:
        page:
          minimum: 0
          type: "integer"
          format: "int32"
        size:
          minimum: 1
          type: "integer"
          format: "int32"
        sort:
          type: "array"
          items:
            type: "string"
    PageUserDto:
      type: "object"
      properties:
        totalPages:
          type: "integer"
          format: "int32"
        totalElements:
          type: "integer"
          format: "int64"
        first:
          type: "boolean"
        last:
          type: "boolean"
        size:
          type: "integer"
          format: "int32"
        content:
          type: "array"
          items:
            $ref: "#/components/schemas/UserDto"
        number:
          type: "integer"
          format: "int32"
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: "integer"
          format: "int32"
        empty:
          type: "boolean"
    PageableObject:
      type: "object"
      properties:
        offset:
          type: "integer"
          format: "int64"
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: "boolean"
        unpaged:
          type: "boolean"
        pageNumber:
          type: "integer"
          format: "int32"
        pageSize:
          type: "integer"
          format: "int32"
    SortObject:
      type: "object"
      properties:
        empty:
          type: "boolean"
        unsorted:
          type: "boolean"
        sorted:
          type: "boolean"
    UserDto:
      required:
      - "firstName"
      - "lastName"
      - "middleName"
      - "phoneNumber"
      type: "object"
      properties:
        id:
          type: "string"
        phoneNumber:
          type: "string"
        firstName:
          type: "string"
        middleName:
          type: "string"
        lastName:
          type: "string"
        isActive:
          type: "boolean"
        cards:
          type: "array"
          items:
            $ref: "#/components/schemas/CardDto"
        fullName:
          type: "string"
...
  securitySchemes:
    Bearer Authentication:
      type: "http"
      description: "Enter JWT token in format: Bearer {token}"
      scheme: "bearer"
      bearerFormat: "JWT"